Change readCommand method from Duke to Parser
